Working paging and PM allocator

+++ include/kernel/hw/interrupt_handlers.h
@@ -4,7 +4,7 @@
 #include <kernel/hw/cpu.h>
 
 typedef void 
-(*exception_handler_t)(int exception, const struct cpu_context *context);
+(*exception_handler_t)(uint64_t exception, const struct cpu_context *context);
 	
 	
 typedef void 
@@ -21,6 +21,6 @@ exception_set_handler(int exception, exception_handler_t handler);
 
 
 void
-default_exception_handler(int exception, const struct cpu_context *context);
+default_exception_handler(uint64_t exception, const struct cpu_context *context);
 
 #endif // _HW_INTERRUPT_HANDLERS_H_
